Refactor snr supporting code into base ImageSource, breakout shared utils/mask
