+++ src/aspire/source/image.py
@@ -17,6 +17,7 @@ from aspire.image.xform import (
     Multiply,
     Pipeline,
 )
+from aspire.noise import AnisotropicNoiseEstimator
 from aspire.operators import (
     CTFFilter,
     IdentityFilter,
@@ -24,7 +25,7 @@ from aspire.operators import (
     PowerFilter,
 )
 from aspire.storage import MrcStats, StarFile
-from aspire.utils import Rotation, grid_2d, trange
+from aspire.utils import Rotation, grid_2d, ratio_to_decibel, support_mask, trange
 
 logger = logging.getLogger(__name__)
 
@@ -146,6 +147,8 @@ class ImageSource(ABC):
 
         # Instantiate the accessor for the `images` property
         self._img_accessor = _ImageAccessor(self._images, self.n)
+        # For base ImageSource, signal is estimated from `images`
+        self._signal_images = self.images
 
         logger.info(f"Creating {self.__class__.__name__} with {len(self)} images.")
 
@@ -790,6 +793,209 @@ class ImageSource(ABC):
                 )
                 im = self.images[i_start:i_end]
                 im.save(mrcs_filepath, overwrite=overwrite)
+
+    def estimate_signal_mean(self, sample_n=None, support_radius=None, batch_size=512):
+        """
+        Estimate the signal mean of `sample_n` projections.
+
+        :param sample_n: Number of images used for estimate.
+            Defaults to all images in source.
+        :param support_radius: Pixel radius used for masking signal support.
+            Default of None will compute inscribed circle, `self.L // 2`.
+        :param batch_size: Images per batch, defaults 512.
+        :returns: Estimated signal mean
+        """
+
+        if sample_n is None:
+            sample_n = self.n
+
+        if sample_n > self.n:
+            logger.warning(
+                f"`estimate_signal_mean` sample_n > Source.n: {sample_n} > {self.n}."
+                f" Accuracy may be impaired, settting sample_n=self.n={self.n}"
+            )
+            sample_n = self.n
+
+        mask = support_mask(self.L, support_radius, dtype=self.dtype)
+
+        # mean is estimated batch-wise, compare with numpy
+        # Note, for simulation we are implicitly assuming taking `sample_n` is random,
+        #   but this does not need to be the case.  We can add a `random_shuffle` param.
+        first_moment = 0.0
+        _denom = sample_n * np.sum(mask)
+        for i in trange(0, sample_n, batch_size):
+            # Gather this batch of images and mask off area outside support_radius
+            images_masked = self._signal_images[i : i + batch_size].asnumpy()[..., mask]
+            # Accumulate first and second moments
+            first_moment += np.sum(images_masked) / _denom
+
+        logger.info(f"Source estimated signal mean: {first_moment}")
+
+        return first_moment
+
+    def estimate_signal_var(self, sample_n=None, support_radius=None, batch_size=512):
+        """
+        Estimate the signal variance of `sample_n` projections.
+
+        :param sample_n: Number of images used for estimate.
+            Defaults to all images in source.
+        :param support_radius: Pixel radius used for masking signal support.
+            Default of None will compute inscribed circle, `self.L // 2`.
+        :param batch_size: Images per batch, defaults 512.
+        :returns: Estimated signal variance.
+        """
+
+        if sample_n is None:
+            sample_n = self.n
+
+        if sample_n > self.n:
+            logger.warning(
+                f"`estimate_signal_var` sample_n > Source.n: {sample_n} > {self.n}."
+                f" Accuracy may be impaired, settting sample_n=self.n={self.n}"
+            )
+            sample_n = self.n
+
+        mask = support_mask(self.L, support_radius, dtype=self.dtype)
+
+        # Var is estimated batch-wise, compare with numpy
+        # np_estimated_var = np.var(self._signal_images[:sample_n].asnumpy()[..., mask])
+        # Note, for simulation we are implicitly assuming taking `sample_n` is random,
+        #   but this does not need to be the case.  We can add a `random_shuffle` param.
+        first_moment = 0.0
+        second_moment = 0.0
+        _denom = sample_n * np.sum(mask)
+        for i in trange(0, sample_n, batch_size):
+            # Gather this batch of images and mask off area outside support_radius
+            images_masked = self._signal_images[i : i + batch_size].asnumpy()[..., mask]
+            # Accumulate first and second moments
+            first_moment += np.sum(images_masked) / _denom
+            second_moment += np.sum(images_masked**2) / _denom
+
+        # E[X**2] - E[X]**2
+        estimated_var = second_moment - first_moment**2
+        logger.info(f"Source estimated signal var: {estimated_var}")
+
+        return estimated_var
+
+    def estimate_signal_power(
+        self,
+        sample_n=None,
+        support_radius=None,
+        batch_size=512,
+        signal_power_method="estimate_signal_mean",
+    ):
+        """
+        Estimate the signal energy of `sample_n` projections using prescribed method.
+
+        :param sample_n: Number of images used for estimate.
+            Defaults to all images in source.
+        :param support_radius: Pixel radius used for masking signal support.
+            Default of None will compute inscribed circle, `self.L // 2`.
+        :param batch_size: Images per batch, defaults 512.
+        :param signal_power_method: Method used for computing signal energy.
+           Defaults to mean via `estimate_signal_mean`.
+           Can use variance method via `estimate_signal_var`.
+
+        :returns: Estimated signal variance.
+        """
+
+        try:
+            signal_estimate_method = getattr(self, signal_power_method)
+        except AttributeError:
+            raise ValueError(
+                f"Cannot find signal_power_method={signal_power_method}."
+                "  Try the default 'estimate_signal_mean' or 'estimate_signal_var'"
+            )
+
+        signal_power = signal_estimate_method(
+            sample_n=sample_n, support_radius=support_radius, batch_size=batch_size
+        )
+        if signal_estimate_method == "estimate_signal_mean":
+            signal_power = signal_power**2  # mean**2
+
+        return signal_power
+
+    def estimate_noise_power(
+        self,
+        sample_n=None,
+        support_radius=None,
+        batch_size=512,
+        noise_estimator=AnisotropicNoiseEstimator,
+    ):
+        """
+        Estimate the noise energy of `sample_n` images using prescribed estimator.
+
+        :param sample_n: Number of images used for estimate.
+            Defaults to all images in source.
+        :param support_radius: Pixel radius used for masking signal support.
+            Default of None will compute inscribed circle, `self.L // 2`.
+        :param batch_size: Images per batch, defaults 512.
+        :param noise_estimator: Method used for estimating noise.
+           Defaults to AnisotropicNoiseEstimator.
+
+        :returns: Estimated noise energy (variance).
+        """
+
+        est = noise_estimator(src=self, bgRadius=support_radius, batchSize=batch_size)
+
+        return est.estimate()
+
+    def estimate_snr(
+        self,
+        sample_n=None,
+        support_radius=None,
+        batch_size=512,
+        noise_power=None,
+        signal_power_method="estimate_signal_mean",
+        units=None,
+    ):
+        """
+        Estimate the SNR of the simulated data set using
+        estimated signal variance / noise variance.
+
+        :param sample_n: Number of images used for estimate.
+            Defaults to all images in source.
+        :param support_radius: Pixel radius used for masking signal support.
+            Default of None will compute inscribed circle, `self.L // 2`.
+        :param batch_size: Images per batch, defaults 512.
+        :param signal_power_method: Method used for computing signal energy.
+           Defaults to mean via `estimate_signal_mean`.
+           Can use variance method via `estimate_signal_var`.
+        :param units: Optionally, convert from default ratio to log scale (`dB`).
+        :returns: Estimated signal to noise ratio.
+        """
+
+        if sample_n is None:
+            sample_n = self.n
+
+        if sample_n > self.n:
+            logger.warning(
+                f"`estimate_snr` sample_n > Source.n: {sample_n} > {self.n}."
+                f" Accuracy may be impaired, settting sample_n=self.n={self.n}"
+            )
+            sample_n = self.n
+
+        if noise_power is None:
+            noise_power = self.estimate_noise_power()
+
+        signal_power = self.estimate_signal_power(
+            sample_n=sample_n,
+            support_radius=support_radius,
+            batch_size=batch_size,
+            signal_power_method=signal_power_method,
+        )
+
+        # For `estimate_signal_mean` we yield: signal_mean**2  / noise_variance
+        #     `estimate_signal_var`   we yield: signal_variance / noise_variance
+        snr = signal_power / noise_power
+
+        # Perform any unit conversion
+        if units == "dB":
+            snr = ratio_to_decibel(snr)
+        elif units is not None:
+            raise ValueError("Units should be `None` or `dB`.")
+
+        return snr
 
 
 class ArrayImageSource(ImageSource):
